+++ src/GRA.Data/Model/VendorCode.cs
@@ -36,5 +36,8 @@ namespace GRA.Data.Model
         [MaxLength(512)]
         public string TrackingNumber { get; set; }
         public DateTime? EmailSentAt { get; set; }
+
+        public bool? IsDamaged { get; set; }
+        public bool? IsMissing { get; set; }
     }
 }