+++ src/GRA.Domain.Model/PackingSlipSummary.cs
@@ -10,7 +10,9 @@ namespace GRA.Domain.Model
         public VendorCodePackingSlip VendorCodePackingSlip { get; set; }
         public bool CanViewDetails { get; set; }
         public string ReceivedBy { get; set; }
-        public bool CanReceivePackingSlips { get; set; }
+        public bool CanBeReceived { get; set; }
         public IEnumerable<string> TrackingNumbers { get; set; }
+        public IEnumerable<int> DamagedItems { get; set; }
+        public IEnumerable<int> MissingItems { get; set; }
     }
 }